Abstract metrics selectors into separate files
